api/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Celery
redis
requests
openai
python-dotenv
